+++ packages/DataTable/stories/DataTable.stories.ts
@@ -8,6 +8,7 @@ import RealWorld from "./examples/RealWorld";
 import RenderRow from "./examples/RenderRow";
 import Showcase from "./examples/Showcase";
 import WithCollapsible from "./examples/WithCollapsible";
+import DataChange from "./examples/DataChange";
 
 export default {
   title: getStoryName("DataTable"),
@@ -29,3 +30,5 @@ export const realWorld = RealWorld;
 export const performance = Performance;
 
 export const withCollapsible = WithCollapsible;
+
+export const dataChange = DataChange;