+++ packages/helpers/src/dom/getActiveElement.js
@@ -1,0 +1,8 @@
+export default function getActiveElement(rootElement = document) {
+  const activeElement = rootElement.activeElement;
+  if (!activeElement) return null;
+
+  const shadowRoot = activeElement.shadowRoot;
+
+  return shadowRoot ? getActiveElement(shadowRoot) : activeElement;
+}