+++ packages/helpers/src/index.js
@@ -5,6 +5,7 @@ export { default as extractChildrenProps } from "./extractChildrenProps";
 export { default as isDevEnv } from "./isDevEnv";
 
 // dom
+export { default as getActiveElement } from "./dom/getActiveElement";
 export { default as isElementContainsFocus } from "./dom/isElementContainsFocus";
 
 // components