+++ packages/ActionBar/CHANGELOG.md
@@ -1,5 +1,34 @@
 # Changelog
 
+## 5.0.4-next.1
+
+### Patch Changes
+
+- Updated dependencies [0dcf307]
+- Updated dependencies [0dcf307]
+  - @paprika/popover@2.0.4-next.1
+  - @paprika/helpers@2.3.0-next.0
+  - @paprika/button@1.1.12-next.1
+  - @paprika/input@4.0.18-next.1
+  - @paprika/select@2.0.12-next.1
+  - @paprika/sortable@1.2.12-next.1
+  - @paprika/switch@1.0.17-next.1
+
+## 5.0.4-next.0
+
+### Patch Changes
+
+- Updated dependencies [0bf34d9]
+  - @paprika/tokens@1.1.5-next.0
+  - @paprika/button@1.1.12-next.0
+  - @paprika/icon@2.1.7-next.0
+  - @paprika/input@4.0.18-next.0
+  - @paprika/popover@2.0.4-next.0
+  - @paprika/select@2.0.12-next.0
+  - @paprika/sortable@1.2.12-next.0
+  - @paprika/stylers@1.1.6-next.0
+  - @paprika/switch@1.0.17-next.0
+
 ## 5.0.3
 
 ### Patch Changes