+++ packages/Card/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 2.0.2-next.1
+
+### Patch Changes
+
+- Updated dependencies [0dcf307]
+- Updated dependencies [0dcf307]
+  - @paprika/helpers@2.3.0-next.0
+
+## 2.0.2-next.0
+
+### Patch Changes
+
+- Updated dependencies [0bf34d9]
+  - @paprika/tokens@1.1.5-next.0
+  - @paprika/stylers@1.1.6-next.0
+
 ## 2.0.1
 
 ### Patch Changes