+++ packages/CollapsibleCard/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## 2.0.21-next.1
+
+### Patch Changes
+
+- Updated dependencies [0dcf307]
+- Updated dependencies [0dcf307]
+  - @paprika/helpers@2.3.0-next.0
+  - @paprika/button@1.1.12-next.1
+  - @paprika/counter@1.1.5-next.1
+
+## 2.0.21-next.0
+
+### Patch Changes
+
+- Updated dependencies [0bf34d9]
+  - @paprika/tokens@1.1.5-next.0
+  - @paprika/button@1.1.12-next.0
+  - @paprika/counter@1.1.5-next.0
+  - @paprika/heading@1.0.14-next.0
+  - @paprika/icon@2.1.7-next.0
+  - @paprika/stylers@1.1.6-next.0
+
 ## 2.0.20
 
 ### Patch Changes