+++ packages/DateInput/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 1.0.28-next.1
+
+### Patch Changes
+
+- Updated dependencies [0dcf307]
+- Updated dependencies [0dcf307]
+  - @paprika/helpers@2.3.0-next.0
+  - @paprika/input@4.0.18-next.1
+
+## 1.0.28-next.0
+
+### Patch Changes
+
+- @paprika/icon@2.1.7-next.0
+- @paprika/input@4.0.18-next.0
+
 ## 1.0.27
 
 ### Patch Changes