+++ packages/ListBox/CHANGELOG.md
@@ -1,5 +1,34 @@
 # Changelog
 
+## 4.0.6-next.1
+
+### Patch Changes
+
+- 0dcf307: Refactored to use `getActiveElement` from `@paprika/helpers` to support web components.
+- Updated dependencies [0dcf307]
+- Updated dependencies [0dcf307]
+  - @paprika/popover@2.0.4-next.1
+  - @paprika/helpers@2.3.0-next.0
+  - @paprika/button@1.1.12-next.1
+  - @paprika/checkbox@1.0.22-next.1
+  - @paprika/input@4.0.18-next.1
+  - @paprika/toast@1.1.20-next.1
+
+## 4.0.6-next.0
+
+### Patch Changes
+
+- Updated dependencies [0bf34d9]
+  - @paprika/tokens@1.1.5-next.0
+  - @paprika/button@1.1.12-next.0
+  - @paprika/checkbox@1.0.22-next.0
+  - @paprika/icon@2.1.7-next.0
+  - @paprika/input@4.0.18-next.0
+  - @paprika/popover@2.0.4-next.0
+  - @paprika/raw-button@1.0.12-next.0
+  - @paprika/stylers@1.1.6-next.0
+  - @paprika/toast@1.1.20-next.0
+
 ## 4.0.5
 
 ### Patch Changes