+++ packages/ListBoxWithTags/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## 3.0.6-next.1
+
+### Patch Changes
+
+- Updated dependencies [0dcf307]
+- Updated dependencies [0dcf307]
+  - @paprika/list-box@4.0.6-next.1
+  - @paprika/helpers@2.3.0-next.0
+  - @paprika/tag@0.0.21-next.1
+
+## 3.0.6-next.0
+
+### Patch Changes
+
+- Updated dependencies [0bf34d9]
+  - @paprika/tokens@1.1.5-next.0
+  - @paprika/icon@2.1.7-next.0
+  - @paprika/list-box@4.0.6-next.0
+  - @paprika/raw-button@1.0.12-next.0
+  - @paprika/stylers@1.1.6-next.0
+  - @paprika/tag@0.0.21-next.0
+
 ## 3.0.5
 
 ### Patch Changes